+++ frontend/src/components/Common/BaseSection/Type.ts
@@ -1,0 +1,4 @@
+
+export interface BaseSectionProps {
+  noHFull?:boolean
+};
